requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.2.5
Django==4.1
django-allauth==0.50.0
django-extensions==3.2.1
django-filter==22.1
django-oauth-toolkit==2.1.0
django-phonenumber-field==7.0.0
django-redis==5.2.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
ecdsa==0.18.0
idna==3.3
importlib-metadata==4.12.0
jwcrypto==1.3.1
Markdown==3.4.1
oauthlib==3.2.0
packaging==21.3
phonenumbers==8.12.56
Pillow==9.2.0
psycopg2==2.9.3
# psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-decouple==3.6
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
wrapt==1.14.1
zipp==3.8.1
